moved pad functions to FixedLengthPadWriter class
