+++ RecordParser/BuilderWrite/SpanExpressionHelper.cs
@@ -147,31 +147,5 @@ namespace RecordParser.BuilderWrite
 
         public static Expression Slice(Expression span, int start, Expression length) =>
             Expression.Call(span, "Slice", Type.EmptyTypes, Expression.Constant(start), length);
-
-        public static Span<char> PadRight(Span<char> span, Span<char> total, char paddingChar)
-        {
-            var diff = total.Length - span.Length;
-            if (diff <= 0) return span;
-
-            span.CopyTo(total);
-
-            for (int i = span.Length, totalWidth = total.Length; i < totalWidth; i++)
-                total[i] = paddingChar;
-
-            return total;
-        }
-
-        public static Span<char> PadLeft(Span<char> span, Span<char> total, char paddingChar)
-        {
-            var diff = total.Length - span.Length;
-            if (diff <= 0) return span;
-
-            span.CopyTo(total.Slice(diff));
-
-            for (var i = 0; i < diff; i++)
-                total[i] = paddingChar;
-
-            return total;
-        }
     }
 }
